int to long

+++ src/lesson8/HW8/Adder.java
@@ -2,7 +2,7 @@ package lesson8.HW8;
 
 public class Adder extends Arithmetic {
 
-    public int add (int a, int b) {
+    public long add (int a, int b) {
         return (a+b);
     }
 }
